Rewrite to use Flask, support Python 3
