+++ .envrc
@@ -1,0 +1,2 @@
+use nix;
+layout python3;
