+++ .gcloudignore
@@ -1,0 +1,6 @@
+.git
+.DS_Store
+*.pyc
+*.nix
+.envrc
+#!include:.gitignore
