+++ requirements.txt
@@ -1,0 +1,3 @@
+appengine-python-standard>=1.0.0
+Flask>=2.2.2
+Jinja2>=3.1.2
